+++ src/main/java/com/ll/netmong/base/exceptionhandler/GlobalExceptionHandler.java
@@ -1,0 +1,33 @@
+package com.ll.netmong.base.exceptionhandler;
+
+import com.ll.netmong.common.RsData;
+import com.ll.netmong.domain.member.exception.NotMatchPasswordException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import javax.security.auth.login.AccountNotFoundException;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(AccountNotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public RsData handleAccountNotFound(AccountNotFoundException e) {
+        return RsData.failOf(e.getMessage());
+    }
+
+    @ExceptionHandler(NotMatchPasswordException.class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    public RsData handleNotMatchPassword(NotMatchPasswordException e) {
+        return RsData.failOf(e.getMessage());
+    }
+
+    // 그 외 예외
+    @ExceptionHandler(Exception.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public RsData handleException(Exception e) {
+        return RsData.failOf("Unexpected error");
+    }
+}
