+++ src/main/java/com/ll/netmong/base/jwt/TokenDto.java
@@ -1,0 +1,16 @@
+package com.ll.netmong.base.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class TokenDto {
+
+    private String access_token;
+    private String refresh_token;
+}
