+++ src/main/java/com/ll/netmong/domain/member/dto/LoginDto.java
@@ -1,0 +1,11 @@
+package com.ll.netmong.domain.member.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginDto {
+    String username;
+    String password;
+}
