+++ src/main/java/com/ll/netmong/domain/member/dto/UsernameRequest.java
@@ -1,0 +1,10 @@
+package com.ll.netmong.domain.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UsernameRequest {
+    String username;
+}
