--- src/main/java/com/ll/netmong/jwt/JwtSecurityConfig.java
@@ -1,19 +1,0 @@
-package com.ll.netmong.jwt;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@RequiredArgsConstructor
-public class JwtSecurityConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-    private final TokenProvider tokenProvider;
-
-    @Override
-    public void configure(HttpSecurity http) {
-        http.addFilterBefore(
-            new JwtFilter(tokenProvider),  UsernamePasswordAuthenticationFilter.class
-        );
-    }
-}
