--- src/main/java/com/ll/netmong/jwt/TokenDto.java
@@ -1,16 +1,0 @@
-package com.ll.netmong.jwt;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenDto {
-
-    private String access_token;
-    private String refresh_token;
-}
