--- src/main/java/com/ll/netmong/jwt/TokenService.java
@@ -1,30 +1,0 @@
-package com.ll.netmong.jwt;
-
-import com.ll.netmong.member.dto.LoginDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class TokenService {
-
-    private final TokenProvider tokenProvider;
-    private final AuthenticationManagerBuilder authenticationManagerBuilder;
-
-    public TokenDto provideTokenWithLoginDto(LoginDto loginDto){
-        UsernamePasswordAuthenticationToken authenticationToken =
-                new UsernamePasswordAuthenticationToken(
-                        loginDto.getUsername(),
-                        loginDto.getPassword());
-
-        Authentication authentication =
-                authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        return tokenProvider.createToken(authentication);
-    }
-}
