--- src/main/java/com/ll/netmong/member/dto/LoginDto.java
@@ -1,11 +1,0 @@
-package com.ll.netmong.member.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginDto {
-    String username;
-    String password;
-}
