--- src/main/java/com/ll/netmong/member/dto/MemberDto.java
@@ -1,18 +1,0 @@
-package com.ll.netmong.member.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-
-import java.util.Collection;
-
-@Getter
-@Builder(toBuilder = true)
-public class MemberDto {
-    private String username;
-    private String password;
-
-    private String realname;
-    private String email;
-    private Collection<? extends GrantedAuthority> roles;
-}
