--- src/main/java/com/ll/netmong/member/dto/UsernameRequest.java
@@ -1,10 +1,0 @@
-package com.ll.netmong.member.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UsernameRequest {
-    String username;
-}
