--- src/main/java/com/ll/netmong/member/repository/MemberJpaRepository.java
@@ -1,10 +1,0 @@
-package com.ll.netmong.member.repository;
-
-import com.ll.netmong.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberJpaRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByUsername(String username);
-}
